Unit tests and cleanup [stage-1]
