+++ angular.json
@@ -479,7 +479,6 @@
               "src/scripts/template-editor/services/svc-template-editor-factory.js",
               "src/scripts/template-editor/services/svc-template-editor-utils.js",
               "src/scripts/template-editor/services/svc-auto-save-service.js",
-              "src/scripts/template-editor/services/svc-blueprint-factory.js",
               "src/scripts/template-editor/services/svc-financial-license-factory.js",
               "src/scripts/template-editor/services/svc-components-factory.js",
