+++ src/app/ajs-upgraded-providers.spec.ts
@@ -1,6 +1,6 @@
 import {expect} from 'chai';
 
-import { AnalyticsFactory, analyticsFactoryProvider, BlueprintFactory, blueprintFactoryProvider, CanvaTypePicker, canvaTypePickerProvider, TemplateEditorFactory, templateEditorFactoryProvider } from './ajs-upgraded-providers';
+import { AnalyticsFactory, analyticsFactoryProvider, CanvaTypePicker, canvaTypePickerProvider, TemplateEditorFactory, templateEditorFactoryProvider } from './ajs-upgraded-providers';
 
 describe('ajs-upgraded-providers', () => {
 
@@ -33,22 +33,6 @@ describe('ajs-upgraded-providers', () => {
       }
       expect(analyticsFactoryProvider.useFactory($injector)).to.equal('service');
       $injector.get.should.have.been.calledWith('analyticsFactory');
-    });
-  });
-
-  describe('blueprintFactory:', () => {    
-    it('should register provider', () => {
-      expect(blueprintFactoryProvider.provide).to.equal(BlueprintFactory);
-      expect(blueprintFactoryProvider.deps).to.deep.equal(['$injector']);
-      expect(blueprintFactoryProvider.useFactory).to.be.a('function');
-    });
-
-    it('should get AngularJS service', () => {
-      const $injector = {
-        get: sinon.stub().returns('service')
-      }
-      expect(blueprintFactoryProvider.useFactory($injector)).to.equal('service');
-      $injector.get.should.have.been.calledWith('blueprintFactory');
     });
   });
