+++ src/app/ajs-upgraded-providers.ts
@@ -19,17 +19,6 @@ export const analyticsFactoryProvider = {
   deps: ['$injector']
 };
 
-export abstract class BlueprintFactory {
-  [key: string]: any;
-}
-export const blueprintFactoryProvider = {
-  provide: BlueprintFactory,
-  useFactory: function ($injector: any) {
-    return $injector.get('blueprintFactory');
-  },
-  deps: ['$injector']
-};
-
 export abstract class TemplateEditorFactory {
   [key: string]: any;
 }
