+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { UpgradeModule } from '@angular/upgrade/static';
 import { HttpClientModule } from '@angular/common/http';
 import { EditorModule } from './editor/editor.module';
-import { analyticsFactoryProvider, blueprintFactoryProvider, canvaTypePickerProvider, templateEditorFactoryProvider } from './ajs-upgraded-providers';
+import { analyticsFactoryProvider, canvaTypePickerProvider, templateEditorFactoryProvider } from './ajs-upgraded-providers';
 import { TemplateEditorModule } from './template-editor/template-editor.module';
 
 @NgModule({
@@ -20,7 +20,6 @@ import { TemplateEditorModule } from './template-editor/template-editor.module';
   ],
   providers: [
     analyticsFactoryProvider,
-    blueprintFactoryProvider,
     canvaTypePickerProvider,
     templateEditorFactoryProvider
   ],
