+++ src/app/template-editor/services/attribute-data.service.ts
@@ -1,15 +1,16 @@
 import { Injectable } from '@angular/core';
-import { BlueprintFactory, TemplateEditorFactory } from 'src/app/ajs-upgraded-providers';
+import { TemplateEditorFactory } from 'src/app/ajs-upgraded-providers';
 import * as angular from 'angular';
 import * as _ from 'lodash';
 import { downgradeInjectable } from '@angular/upgrade/static';
+import { BlueprintService } from './blueprint.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AttributeDataService {
 
-  constructor(private blueprintFactory: BlueprintFactory, private templateEditorFactory: TemplateEditorFactory) {}
+  constructor(private blueprintFactory: BlueprintService, private templateEditorFactory: TemplateEditorFactory) {}
 
   getBlueprintData(componentId, attributeKey?) {
     return this.blueprintFactory.getBlueprintData(componentId, attributeKey);
